requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
dialogflow==1.1.0
feedparser==6.0.2
Flask==2.0.1
fuzzywuzzy==0.18.0
gnewsclient==1.12
google-api-core==1.30.0
google-auth==1.30.2
googleapis-common-protos==1.53.0
grpcio==1.38.0
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==20.9
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-telegram-bot==13.6
pytz==2021.1
requests==2.25.1
rsa==4.7.2
sgmllib3k==1.0.0
six==1.16.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.5
Werkzeug==2.0.1
